requirements.txt
beautifulsoup4
black
pdfminer.six